requirements.txt
numpy
azure-storage-blob
azure-identity
psycopg
Pillow
pydantic
